Update to 24w06a

+++ src/main/java/io/github/drakonkinst/worldsinger/entity/ThrownSporeBottleEntity.java
@@ -117,8 +117,8 @@ public class ThrownSporeBottleEntity extends ThrownItemEntity implements FlyingI
     }
 
     @Override
-    protected float getGravity() {
-        return 0.05f;
+    protected double getGravity() {
+        return 0.05;
     }
 
     @Override
